Extract ModuleTypeResolver, PackageResolveExecutor, and ShowDependenciesParser into nested types of PackageResolver
